# docs/swagger.yaml
basePath: /
definitions:
  api_handler.AuthSuccessResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      message:
        example: Login successful
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      status:
        example: 200
        type: integer
    type: object
  api_handler.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    type: object
...
  api_handler.ErrorResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/api_handler.FieldError'
        type: array
      message:
        example: Invalid request
        type: string
      status:
        example: 400
        type: integer
    type: object
...
  api_handler.FieldError:
    properties:
      field:
        example: email
        type: string
      message:
        example: Email is required
        type: string
    type: object
  api_handler.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  api_handler.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  api_handler.MessageResponse:
    properties:
      message:
        example: Operation completed successfully
        type: string
      status:
        example: 200
        type: integer
    type: object
...
  api_handler.PasswordResetSuccessResponse:
    properties:
      message:
        example: If the email exists, a password reset link has been sent
        type: string
      status:
        example: 200
        type: integer
    type: object
...
  api_handler.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  api_handler.ResetPasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
...
  api_handler.SuccessResponse:
    properties:
      data:
        type: object
      message:
        example: Operation completed successfully
        type: string
      status:
        example: 200
        type: integer
    type: object
  api_handler.TokenVerificationResponse:
    properties:
      expires_at:
        example: "2025-10-20T12:30:00Z"
        type: string
      message:
        example: Token is valid
        type: string
      valid:
        example: true
        type: boolean
    type: object
...
  github_com_edsonmubezi_myapp_internal_organization.Organization:
    properties:
      address:
        minLength: 6
        type: string
      contact_person:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        maxLength: 100
        minLength: 3
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
    required:
    - address
    - contact_person
    - name
    - phone_number
    type: object
...
  github_com_edsonmubezi_myapp_internal_permission.Permission:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      visibility:
        description: '"ALL" | "HQ_ONLY"'
        type: string
    type: object
...
  github_com_edsonmubezi_myapp_internal_role.Role:
    properties:
      id:
        type: integer
      name:
        maxLength: 50
        minLength: 3
        type: string
      organization_id:
        type: integer
      permissions:
        items:
          $ref: '#/definitions/github_com_edsonmubezi_myapp_internal_permission.Permission'
        type: array
    required:
    - name
    - organization_id
    type: object
...
  github_com_edsonmubezi_myapp_internal_user.User:
    properties:
      active_status:
        type: boolean
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      employee_id:
        description: if you added users.employee_id FK
        type: integer
      fullname:
        maxLength: 100
        minLength: 3
        type: string
      id:
        type: integer
      organization:
        $ref: '#/definitions/github_com_edsonmubezi_myapp_internal_organization.Organization'
      organization_id:
        type: integer
      password:
        minLength: 6
        type: string
      photo:
        type: string
      role:
        $ref: '#/definitions/github_com_edsonmubezi_myapp_internal_role.Role'
      role_id:
        type: integer
      updated_at:
        type: string
    required:
    - email
    - fullname
    - password
    - role_id
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@ekklesiahr.com
    name: EkklesiaHR API Support
    url: https://github.com/edsonmubezi/ekklesiahr
  description: Production-ready HR Management System API with comprehensive employee,
    payroll, leave management, and organization administration. Supports multi-tenant
    architecture with role-based access control.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: EkklesiaHR API
  version: 1.0.1
paths:
...
  /api/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Sends a password reset email to the user. Returns the same success
        message regardless of whether the email exists (security best practice).
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_handler.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Reset email sent (or email not found)
          schema:
            $ref: '#/definitions/api_handler.PasswordResetSuccessResponse'
        "400":
          description: Invalid email format
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
      summary: Request Password Reset
      tags:
      - Authentication
  /api/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Resets user password using a valid reset token. Token can only
        be used once and expires after 30 minutes.
      parameters:
      - description: Reset token and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_handler.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successful
          schema:
            $ref: '#/definitions/api_handler.MessageResponse'
        "400":
          description: Invalid token or weak password
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
      summary: Reset Password
      tags:
      - Authentication
  /api/auth/verify-reset-token/{token}:
    get:
      description: Validates a password reset token without consuming it. Use this
        to check token validity before showing password reset form.
      parameters:
      - description: Reset token
        example: '"base64-encoded-token-string"'
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Token is valid
          schema:
            $ref: '#/definitions/api_handler.TokenVerificationResponse'
        "400":
          description: Invalid or expired token
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
      summary: Verify Reset Token
      tags:
      - Authentication
...
  /api/change-password:
    put:
      consumes:
      - application/json
      description: Changes the password for the authenticated user. Requires current
        password verification.
      parameters:
      - description: Current and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_handler.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
          schema:
            $ref: '#/definitions/api_handler.MessageResponse'
        "400":
          description: Invalid request or weak password
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "401":
          description: Unauthorized or incorrect current password
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change Password
      tags:
      - Authentication
...
  /api/login:
    post:
      consumes:
      - application/json
      description: Authenticates user with email and password, returns access and
        refresh tokens. Supports multi-tenant architecture.
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/api_handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with tokens
          schema:
            $ref: '#/definitions/api_handler.AuthSuccessResponse'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "403":
          description: Account is deactivated
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
      summary: User Login
      tags:
      - Authentication
  /api/logout:
    post:
      description: Logs out the authenticated user by revoking their current access
        token. Forces re-authentication for subsequent requests.
      produces:
      - application/json
      responses:
        "200":
          description: Logged out successfully
          schema:
            $ref: '#/definitions/api_handler.MessageResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: User Logout
      tags:
      - Authentication
...
  /api/refresh:
    post:
      consumes:
      - application/json
      description: Generates new access and refresh tokens using a valid refresh token.
        Use this when access token expires.
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_handler.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: New tokens generated successfully
          schema:
            $ref: '#/definitions/api_handler.AuthSuccessResponse'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
      summary: Refresh Access Token
      tags:
      - Authentication
...
  /api/users/me:
    get:
      description: Retrieves the currently authenticated user's details including
        role and organization. Returns user with encrypted ID for security.
      produces:
      - application/json
      responses:
        "200":
          description: Authenticated user fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/api_handler.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_edsonmubezi_myapp_internal_user.User'
              type: object
        "401":
          description: Unauthorized - user ID missing
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "500":
          description: Failed to retrieve user
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Authenticated User
      tags:
      - Users
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: 'Provide your JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"